requirement.txt
adjustText==0.8
aeppl==0.0.35
aesara==2.8.2
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.9.3
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arviz==0.12.1
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
bambi==0.9.1
beautifulsoup4==4.11.1
bioinfokit==2.1.0
black==23.1.0
bleach==5.0.1
Boruta==0.3
cachetools==5.2.0
category-encoders==2.6.0
certifi==2022.6.15
cffi==1.15.1
cftime==1.6.1
charset-normalizer==2.1.1
click==8.1.3
clikit==0.6.2
cloudpickle==2.2.0
cmaes==0.9.1
colorlog==6.7.0
cons==0.4.5
crashtest==0.3.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
deprecat==2.1.1
dill==0.3.5.1
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
etuples==0.3.8
exceptiongroup==1.1.1
executing==1.0.0
fastjsonschema==2.16.1
fastprogress==1.0.3
filelock==3.8.0
fonttools==4.37.1
formulae==0.3.4
frozenlist==1.3.1
future==0.18.3
greenlet==2.0.2
h2o==3.40.0.2
httpstan==4.8.2
idna==3.3
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==2.0.0
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.15.0
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyterlab==3.4.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.15.1
kiwisolver==1.4.4
lightgbm==3.3.5
llvmlite==0.40.0
lofo-importance==0.3.3
logical-unification==0.4.5
lxml==4.9.1
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.17.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
miniKanren==1.0.3
mistune==2.0.4
multidict==6.0.2
multipledispatch==0.6.0
mypy-extensions==1.0.0
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
netCDF4==1.6.0
networkx==3.0
notebook==6.4.12
notebook-shim==0.1.0
numba==0.57.0
numpy==1.21.6
openpyxl==3.0.10
optuna==3.1.0
packaging==23.0
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pastel==0.2.1
pathspec==0.11.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydot==1.4.2
Pygments==2.13.0
pylev==1.4.0
pymc==4.1.7
pymc3==3.11.5
pyparsing==3.0.9
pyrsistent==0.18.1
pysimdjson==3.2.0
pystan==3.5.0
pytest==7.3.1
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.7.3
seaborn==0.12.0
seaborn-qqplot==0.5.0
semver==2.13.0
Send2Trash==1.8.0
shap==0.41.0
shap-selection==0.1.6
six==1.16.0
skdag==0.0.6
sklearn==0.0
slicer==0.0.7
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.3
stack-data==0.5.0
stackeddag==0.3.4
statsmodels==0.13.2
tabulate==0.8.10
terminado==0.15.0
textwrap3==0.9.2
Theano-PyMC==1.1.2
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webargs==8.2.0
webencodings==0.5.1
websocket-client==1.4.1
widgetsnbextension==4.0.5
wrapt==1.14.1
xarray==2022.6.0
xarray-einstats==0.3.0
xgboost==1.7.4
yarg==0.1.9
yarl==1.8.1
zipp==3.8.1
